# forum-project/docs/auth-openapi-spec.yaml
openapi: 3.0.0
info:
  title: 论坛系统用户认证API
  description: |
    论坛系统的用户认证模块API接口文档。
    提供用户注册、登录、登出、令牌管理等功能。
  version: 1.0.0
  contact:
    name: API支持团队
    email: api-support@forum.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api
    description: 开发环境
  - url: https://api.forum.com/v1
    description: 生产环境
...
paths:
...
  /auth/refresh:
    post:
      tags:
        - 令牌
      summary: 刷新令牌
      description: 使用刷新令牌获取新的访问令牌
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: 令牌刷新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: 刷新令牌无效或过期
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ==================== 用户信息接口 ====================
  /auth/me:
    get:
      tags:
        - 用户
      summary: 获取当前用户信息
      description: 获取当前认证用户的详细信息
      operationId: getCurrentUser
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - 用户
      summary: 更新用户信息
      description: 更新当前用户的个人资料
      operationId: updateUserProfile
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ==================== 密码管理接口 ====================
...
components:
  schemas:
...
    RefreshTokenRequest:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
          description: 刷新令牌
...
    UpdateProfileRequest:
      type: object
      properties:
        displayName:
          type: string
          maxLength: 50
          example: John Smith
          description: 显示名称
        avatar:
          type: string
          format: uri
          example: https://example.com/avatar.jpg
          description: 头像URL
        bio:
          type: string
          maxLength: 500
          example: 热爱编程的开发者
          description: 个人简介
        currentPassword:
          type: string
          example: CurrentPassword123
          description: 当前密码 (修改密码时需要)
        newPassword:
          type: string
          minLength: 8
          example: NewPassword123
          description: 新密码 (修改密码时需要)
        confirmNewPassword:
          type: string
          example: NewPassword123
          description: 确认新密码 (修改密码时需要)

    # ==================== 响应体 ====================
    SuccessResponse:
      type: object
      required:
        - success
        - message
        - timestamp
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: 操作成功
        data:
          type: object
          nullable: true
        timestamp:
          type: string
          format: date-time
          example: 2024-01-01T12:00:00.000Z
    ErrorResponse:
      type: object
      required:
        - success
        - message
        - error
        - timestamp
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: 请求参数验证失败
        error:
          type: object
          required:
            - code
            - details
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            details:
              type: string
              example: 用户名格式不正确
            field:
              type: string
              example: username
        timestamp:
          type: string
          format: date-time
          example: 2024-01-01T12:00:00.000Z
...
    TokenResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/TokenInfo'
    UserResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/User'
...
    TokenInfo:
      type: object
      required:
        - token
        - refreshToken
        - expiresIn
        - tokenType
      properties:
        token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
          description: 访问令牌
        refreshToken:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
          description: 刷新令牌
        expiresIn:
          type: integer
          example: 3600
          description: 令牌有效期 (秒)
        tokenType:
          type: string
          enum: [Bearer]
          example: Bearer
          description: 令牌类型
    User:
      type: object
      required:
        - id
        - username
        - email
        - emailVerified
        - role
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        username:
          type: string
          example: john_doe
        email:
          type: string
          format: email
          example: john@example.com
        displayName:
          type: string
          example: John Doe
        avatar:
          type: string
          format: uri
          example: https://example.com/avatar.jpg
        bio:
          type: string
          example: 热爱编程的开发者
        emailVerified:
          type: boolean
          example: true
        role:
          type: string
          enum: [user, moderator, admin]
          example: user
        status:
          type: string
          enum: [active, suspended, banned, inactive]
          example: active
        createdAt:
          type: string
          format: date-time
          example: 2024-01-01T12:00:00.000Z
        updatedAt:
          type: string
          format: date-time
          example: 2024-01-01T12:00:00.000Z
        lastLoginAt:
          type: string
          format: date-time
          example: 2024-01-01T12:00:00.000Z
        stats:
          $ref: '#/components/schemas/UserStats'
    UserStats:
      type: object
      properties:
        postCount:
          type: integer
          example: 42
          description: 发帖数
        commentCount:
          type: integer
          example: 156
          description: 评论数
        likeCount:
          type: integer
          example: 320
          description: 获赞数
        followerCount:
          type: integer
          example: 89
          description: 粉丝数
        followingCount:
          type: integer
          example: 45
          description: 关注数
        reputation:
          type: integer
          example: 1250
          description: 声望值
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        使用JWT令牌进行认证。
        格式: `Authorization: Bearer <token>`

# ==================== 扩展信息 ====================
...